kernel: rtl8367b: initial support for Realtek switch rtl8367s
From driver point of view no differance between rtl8367b and rtl8367s
if it connected through EXT2 (rgmii only).
So this trivial patch add some identification and initialization only.
SGMII/HSGMII mode for EXT1 is not implemented for the sake of patch clairity.
It may be simple but without test device...

Signed-off-by: Serge Vasilugin <vasilugin@yandex.ru>